+++ src/interpreter/ts/primitive.rs
@@ -19,6 +19,7 @@ impl Interpreter for Primitive {
         Ok(buf.write_all(
             match self {
                 Self::Number(..) => "number",
+                Self::BigInt(..) => "BigInt",
                 Self::String(..) => "string",
                 Self::Boolean(..) => "boolean",
             }