+++ src/nature/defs/primitive.rs
@@ -8,6 +8,7 @@ use quote::{format_ident, quote};
 #[derive(Clone, Debug)]
 pub enum Primitive {
     Number(OriginType),
+    BigInt(OriginType),
     String(OriginType),
     Boolean(OriginType),
 }
@@ -16,6 +17,7 @@ impl TypeTokenStream for Primitive {
     fn type_token_stream(&self) -> Result<TokenStream, E> {
         match self {
             Self::Number(ty) => ty,
+            Self::BigInt(ty) => ty,
             Self::String(ty) => ty,
             Self::Boolean(ty) => ty,
         }
@@ -27,6 +29,7 @@ impl TypeAsString for Primitive {
     fn type_as_string(&self) -> Result<String, E> {
         match self {
             Self::Number(ty) => ty,
+            Self::BigInt(ty) => ty,
             Self::String(ty) => ty,
             Self::Boolean(ty) => ty,
         }