+++ tests/node-bindgen/rs/types.rs
@@ -3,24 +3,24 @@ use tslink::tslink;
 
 #[tslink(snake_case_naming)]
 #[node_bindgen]
-fn types_a(a: i64, b: i32) -> (i64, i32) {
+fn types_a(a: i32, b: i32) -> (i32, i32) {
     (a, b)
 }
 
 #[tslink(snake_case_naming)]
 #[node_bindgen]
-fn types_b(a: Option<i64>) -> Option<i64> {
+fn types_b(a: Option<i32>) -> Option<i32> {
     a
 }
 
 #[tslink(snake_case_naming)]
 #[node_bindgen]
-fn types_c(a: Option<i64>, b: Option<i64>) -> (Option<i64>, Option<i64>) {
+fn types_c(a: Option<i32>, b: Option<i32>) -> (Option<i32>, Option<i32>) {
     (a, b)
 }
 
 #[tslink(snake_case_naming)]
 #[node_bindgen]
-fn types_d(a: Vec<i64>) -> Vec<i64> {
+fn types_d(a: Vec<i32>) -> Vec<i32> {
     a
 }