+++ changelog.md
@@ -1,3 +1,11 @@
+# 0.1.5 (24.20.2024)
+
+## Fixes
+
+- Use BigInt type for u64, u128, usize, i64, i128, isize
+- Correctly parse f16, f32 and f64 types
+- Serialize names like r#field 
+
 # 0.1.4 (23.20.2024)
 
 ## Fixes